// frontend/src/types/Movie.ts
export interface Movie{
    showId: string | null;
    type: string | null;
    title: string | null;
    director: string | null;
    cast: string | null;
    country: string | null;
    releaseYear: number | null;
    rating: string | null;
    duration: string | null;
    description: string | null;
    action: number | null;
    adventure: number | null;
    animeSeriesInternationalTvShows: number | null;
    britishTvShowsDocuseriesInternationalTvShows: number | null;
    children: number | null;
    comedies: number | null;
    comediesDramasInternationalMovies: number | null;
    comediesInternationalMovies: number | null;
    comediesRomanticMovies: number | null;
    crimeTvShowsDocuseries: number | null;
    documentaries: number | null;
    documentariesInternationalMovies: number | null;
    docuseries: number | null;
    dramas: number | null;
    dramasInternationalMovies: number | null;
    dramasRomanticMovies: number | null;
    familyMovies: number | null;
    fantasy: number | null;
    horrorMovies: number | null;
    internationalMoviesThrillers: number | null;
    internationalTvShowsRomanticTvShowsTvDramas: number | null;
    kidsTv: number | null;
    languageTvShows: number | null;
    musicals: number | null;
    natureTv: number | null;
    realityTv: number | null;
    spirituality: number | null;
    tvAction: number | null;
    tvComedies: number | null;
    tvDramas: number | null;
    talkShowsTvComedies: number | null;
    thrillers: number | null;
  }